提交中央仓

+++ xm-biz/src/main/java/com/xunmo/webs/users/controller/UsersController.java
@@ -1,11 +1,11 @@
 package com.xunmo.webs.users.controller;
 
+import cn.hutool.core.collection.CollUtil;
 import cn.hutool.core.date.DateUtil;
 import cn.hutool.core.util.StrUtil;
 import com.xunmo.common.base.BaseController;
 import com.xunmo.common.entity.ResponseEntity;
-import com.xunmo.common.utils.ResponseUtil;
-import com.xunmo.enums.SystemStatus;
+import com.xunmo.common.utils.R;
 import com.xunmo.jimmer.annotation.Db;
 import com.xunmo.jimmer.page.Page;
 import com.xunmo.jimmer.page.PageRequest;
@@ -28,13 +28,17 @@ import org.noear.solon.annotation.Mapping;
 import org.noear.solon.annotation.Post;
 import org.noear.solon.data.annotation.CacheRemove;
 import org.noear.solon.data.annotation.Tran;
-import org.noear.solon.validation.annotation.NotBlank;
 import org.noear.solon.validation.annotation.NotNull;
 import org.noear.solon.validation.annotation.Valid;
 import org.noear.solon.validation.annotation.Validated;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
+
+import static com.xunmo.enums.SystemStatus.IS_SUCCESS;
 
 /**
  * 用户表(Users)表控制层
@@ -69,7 +73,7 @@ public class UsersController extends BaseController {
 	@Mapping("/list")
 //	@Cache(tags = "users", seconds = 10)
 	public ResponseEntity<Page<Users>> list(@Validated UsersQuery query, PageRequest pageRequest) throws Exception {
-		final String userId = query.getUserId();
+		final String usersId = query.getUsersId();
 		final String userName = query.getUserName();
 		final LocalDateTime beginCreateTime = query.getBeginCreateTime();
 		final LocalDateTime endCreateTime = query.getEndCreateTime();
@@ -81,7 +85,7 @@ public class UsersController extends BaseController {
 		final String permissionName = query.getPermissionName();
 		final Page<Users> page = pager(pageRequest).execute(sqlClient.createQuery(TABLE)
 				// 根据 用户id 查询
-				.whereIf(StrUtil.isNotBlank(userId), () -> TABLE.usersId().eq(userId))
+				.whereIf(StrUtil.isNotBlank(usersId), () -> TABLE.usersId().eq(usersId))
 				// 根据 用户名称 模糊查询
 				.whereIf(StrUtil.isNotBlank(userName), () -> TABLE.userName().like(userName))
 				// 根据 组织机构id 查询
@@ -116,45 +120,60 @@ public class UsersController extends BaseController {
 								.roles(RoleFetcher.$.parentRoleId()
 										.roleName()
 										.permissions(PermissionFetcher.$.permissionName())))));
-		return ResponseUtil.genResponse(SystemStatus.IS_SUCCESS, page);
-	}
-
-	/**
-	 * 通过主键查询单条数据
-	 *
-	 * @param id 主键
-	 * @return 单条数据
-	 */
-	@Post
-	@Mapping("/getById")
-	public ResponseEntity<Users> getById(@NotNull @NotBlank String id) throws Exception {
-		return ResponseUtil.genResponse(SystemStatus.IS_SUCCESS, this.sqlClient.findById(Users.class, id));
+		return R.out(IS_SUCCESS, page);
 	}
 
 	/**
 	 * 新增数据
 	 *
-	 * @param input 实体
+	 * @param inputs 实体s
 	 * @return 新增结果
 	 */
 	@Post
-	@Mapping("/add")
-	public ResponseEntity<Users> add(@Validated UsersInput input) throws Exception {
-		final SimpleSaveResult<Users> result = this.sqlClient.getEntities().save(input);
-		return ResponseUtil.genResponse(SystemStatus.IS_SUCCESS, result.getModifiedEntity());
+	@Mapping("/adds")
+	public ResponseEntity<Users> add(@Validated List<UsersInput> inputs) throws Exception {
+		final List<Users> users = inputs.stream()
+				.map(UsersInput::toEntity)
+				.collect(Collectors.toList());
+		final List<SimpleSaveResult<Users>> simpleResults = this.sqlClient.getEntities().batchSave(users).getSimpleResults();
+		final List<Users> usersModifieds = simpleResults.stream()
+				.map(SimpleSaveResult::getModifiedEntity)
+				.collect(Collectors.toList());
+		return R.out(IS_SUCCESS, usersModifieds);
 	}
 
 	/**
 	 * 编辑数据
 	 *
-	 * @param input 实体
+	 * @param inputs 实体s
 	 * @return 编辑结果
 	 */
 	@Post
-	@Mapping("/update")
-	public ResponseEntity<Users> update(@Validated UsersInput input) throws Exception {
-		final SimpleSaveResult<Users> result = this.sqlClient.update(input);
-		return ResponseUtil.genResponse(SystemStatus.IS_SUCCESS, result.getModifiedEntity());
+	@Mapping("/updates")
+	public ResponseEntity<Users> update(@Validated List<UsersInput> inputs) throws Exception {
+		final List<Users> usersModifieds = new ArrayList<>(inputs.size());
+		for (UsersInput input : inputs) {
+			final SimpleSaveResult<Users> result = this.sqlClient.update(input);
+			usersModifieds.add(result.getModifiedEntity());
+		}
+		return R.out(IS_SUCCESS, usersModifieds);
+	}
+
+	/**
+	 * 通过主键查询单条数据
+	 *
+	 * @param ids 主键s
+	 * @return 单条数据
+	 */
+	@Post
+	@Mapping("/getByIds")
+	public ResponseEntity<Users> getByIds(@NotNull List<String> ids) throws Exception {
+		if (CollUtil.isEmpty(ids)) {
+			return R.out(IS_SUCCESS, Collections.emptyList());
+		}
+//		final List<Users> usersList = this.sqlClient.findByIds(Users.class, ids);
+		final List<Users> usersList = usersRepository.findByIds(ids);
+		return R.out(IS_SUCCESS, usersList);
 	}
 
 	/**
@@ -164,11 +183,14 @@ public class UsersController extends BaseController {
 	 * @return 删除是否成功
 	 */
 	@Post
-	@Mapping("/deleteByIds")
+	@Mapping("/delByIds")
 	@Tran
-	public ResponseEntity<Boolean> deleteByIds(List<String> ids) throws Exception {
+	public ResponseEntity<Boolean> deleteByIds(@NotNull List<String> ids) throws Exception {
+		if (CollUtil.isEmpty(ids)) {
+			return R.out(IS_SUCCESS, Collections.emptyList());
+		}
 		final DeleteResult result = this.sqlClient.deleteByIds(Users.class, ids);
-		return ResponseUtil.genResponse(SystemStatus.IS_SUCCESS, result.getTotalAffectedRowCount());
+		return R.out(IS_SUCCESS, result.getTotalAffectedRowCount());
 	}
 
 	/**
@@ -179,21 +201,6 @@ public class UsersController extends BaseController {
 	@CacheRemove(tags = "users")
 	public ResponseEntity<Boolean> exception(PageRequest pageRequest) throws Exception {
 		throw new NullPointerException("主动抛出异常 - 用于测试 " + DateUtil.now());
-	}
-
-	/**
-	 * 分页查询
-	 *
-	 * @param query       筛选条件
-	 * @param pageRequest 分页对象
-	 * @return 查询结果
-	 */
-	@Post
-	@Mapping("/testRepository")
-	public ResponseEntity<Page<Users>> testRepository(@Validated UsersQuery query, PageRequest pageRequest)
-			throws Exception {
-		final long count = usersRepository.count();
-		return ResponseUtil.genResponse(SystemStatus.IS_SUCCESS, count);
 	}
 
 }
